refactoring of create to prevent conflicts with previous pr
