+++ routes/auth.js
@@ -5,6 +5,8 @@ const router = express.Router()
 const authController = require('../controllers/auth')
 const authValidations = require('../middlewares/auth-validations')
 
+router.get('/', authController.getAll)
 router.post('/login', authValidations.login, authController.login)
+router.post('/register', authController.create)
 
 module.exports = router
