--- routes/users.js
@@ -1,10 +1,0 @@
-const express = require('express')
-const usersController = require('../controllers/users')
-const usersValidation = require('../middlewares/users-validation')
-
-const router = express.Router()
-
-router.get('/', usersController.getAll)
-router.post('/register', usersValidation, usersController.create)
-
-module.exports = router
